% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktm.R
\docType{package}
\name{ktm}
\alias{ktm}
\alias{ktm-package}
\title{ktm: Korean Text Mining toolkit using `Seunjeon`, `dplyr`, and other tools}
\description{
This package utilizes various tools for text mining to analyze Korean text.
}
\details{
In this package, every result return in UTF-8 to avoid encoding collision,
which is not ensured in many text mining packages.

This is build for making Korean text mining easy, which means there are possible
performance losses in several funtions. It'll be improved with update, which will
offer connections to other text mining tools.
}
